CONTRIBUTING: Add Diamond Bluff contribution steps
A first crack at adding how to contribute to the Diamond Bluff project.
This is very basic at the moment, with the key concept being mentioning
the [DCO](https://developercertificate.org), which is required for all
commits into the Diamond Bluff repository.

The second thing this commit does is a first crack at setting up a
MAINTAINERS.md file. I've taken this from the list of people in the
Diamond Bluff github project, and whittled it down to people who have
attended meetings regularly. We'll want to trim this back even further
at some point.

Signed-off-by: Kyle Mestery <mestery@mestery.com>